% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.linreg}
\alias{coef.linreg}
\title{Coef method for linreg objects}
\usage{
\method{coef}{linreg}(object)
}
\arguments{
\item{object}{An object of class "linreg".}
}
\value{
A vector of coefficients.
}
\description{
Extract coefficients from an object of the class "linreg".
}
\examples{
 
}
