% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_ridgereg.R
\name{coef.ridgereg}
\alias{coef.ridgereg}
\title{Coef method for ridgereg objects}
\usage{
\method{coef}{ridgereg}(object, ...)
}
\arguments{
\item{object}{An object of class "ridgereg", generated by a call to \code{\link{ridgereg}}.}

\item{...}{additional arguments.}
}
\value{
A vector of coefficients.
}
\description{
Extract coefficients from an object of the class "ridgereg".
}
\examples{
ridgereg_obj <- ridgereg(Petal.Length~Species, datasets::iris)
coef(ridgereg_obj)
}
