% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab4HugoOtto.R
\docType{package}
\name{Lab4HugoOtto}
\alias{Lab4HugoOtto}
\title{Lab4HugoOtto: A package containing the a fuction for linear regression}
\description{
The package provides a function called \code{\link{linreg}}. This function performs linear regression using ordinary least squares. The function
returns an object of the S3 class "linreg" which can be explored with the followin methods:
\itemize{
  \item \code{\link{coef.linreg}}
  \item \code{\link{plot.linreg}}
  \item \code{\link{pred.linreg}}
  \item \code{\link{print.linreg}}
  \item \code{\link{residuals.linreg}}
  \item \code{\link{summary.linreg}}
}
}
