% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.linreg}
\alias{plot.linreg}
\title{Plot method for linreg xs}
\usage{
\method{plot}{linreg}(x, ...)
}
\arguments{
\item{x}{an object of class "linreg", generated by a call to \code{\link{linreg}}.}

\item{...}{additional arguments.}
}
\value{
Two scatter plots. The first one is residuals vs fitted values and the second is the scale-location.
}
\description{
\code{plot} method for xs of the class "linreg".
}
