% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{pred}
\alias{pred}
\title{Predict method for linreg objects}
\usage{
pred(object, ...)
}
\arguments{
\item{object}{An object of class "linreg", generated by a call to \code{\link{linreg}}.}

\item{...}{additional arguments.}
}
\value{
A vector of predicted values.
}
\description{
Predicted values from an object of the class "linreg".
}
