% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_ridgereg.R
\name{predict.ridgereg}
\alias{predict.ridgereg}
\title{Predict method for ridgereg objects}
\usage{
\method{predict}{ridgereg}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{An object of class "ridgereg", generated by a call to \code{\link{ridgereg}}.}

\item{...}{additional arguments.}
}
\value{
A vector of predicted values.
}
\description{
Predicted values from an object of the class "ridgereg".
}
\examples{
ridgereg_obj <- ridgereg(Petal.Length~Species, datasets::iris)
predict(ridgereg_obj)
}
