% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_ridgereg.R
\name{print.ridgereg}
\alias{print.ridgereg}
\title{Print out the estimated coeffiencts}
\usage{
\method{print}{ridgereg}(x, ...)
}
\arguments{
\item{x}{An object of Class "ridgereg", generated by a call to \code{\link{ridgereg}}.}

\item{...}{additional arguments.}
}
\value{
A list of estimated coeffiencts.
}
\description{
Estimated coeffiencts from an x of class "ridgereg".
}
\examples{
ridgereg_obj <- ridgereg(Petal.Length~Species, datasets::iris)
print(ridgereg_obj)
}
