% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals.linreg}
\alias{residuals.linreg}
\title{Residuals method for linreg objects}
\usage{
\method{residuals}{linreg}(object)
}
\arguments{
\item{object}{An object of class "linreg".}
}
\value{
A vector of residuals.
}
\description{
Residuals from an object of the class "linreg".
}
\examples{
  
}
