% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\name{ridgereg}
\alias{ridgereg}
\title{Perform Ridge regression with OLS}
\usage{
ridgereg(formula, data, lambda)
}
\arguments{
\item{formula}{An object of the class \code{\link[stats]{formula}}.}

\item{data}{A data set of the class \code{\link[base]{data.frame}}.}

\item{lambda}{A numeric value.}
}
\value{
The function returns the results of the Ridge regression as an object of class "ridgereg".
}
\description{
Perform Ridge regression with OLS
}
\examples{
ridredge_obj <- ridgereg(Petal.Length~Species, datasets::iris)
}
