% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.linreg}
\alias{summary.linreg}
\alias{print.summary.linreg}
\title{Summary method for linreg objects}
\usage{
\method{summary}{linreg}(object, ...)

\method{print}{summary.linreg}(x, ...)
}
\arguments{
\item{object}{an object of class "linreg", generated by a call to \code{\link{linreg}}.}

\item{x}{an object of class "summary.linreg", generated by a call to \code{summary.linreg}.}
}
\value{
A list of summary statistics from a linear regression model, \code{object}, of the class "linreg".
}
\description{
\code{summary} method for objects of the class .
}
\section{Functions}{
\itemize{
\item \code{print.summary.linreg}: 
}}

