% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_linkoping.R
\name{theme_linkoping}
\alias{theme_linkoping}
\title{ggplot2 theme with colors from the Linkoping University graphical manual}
\usage{
theme_linkoping()
}
\description{
ggplot2 theme with colors from the Linkoping University graphical manual
}
