% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_airport_delays.R
\name{visualize_airport_delays}
\alias{visualize_airport_delays}
\title{Visualize delays for airports in the USA}
\usage{
visualize_airport_delays()
}
\value{
Returns a ggplot object containing a scatterplot showing delays at different airport in the USA
}
\description{
Uses ggplot2 to create a scatterplot showing delays at different airport in the USA
}
\examples{
airport_delays <- visualize_airport_delays()
airport_delays
}
