% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab4HugoOtto.R
\docType{package}
\name{Lab4HugoOtto}
\alias{Lab4HugoOtto}
\title{Lab4HugoOtto: A package containing functions for linear and ridge regression}
\description{
The package provides two functions. The first function is called \code{\link{linreg}}. This function performs linear regression using ordinary least squares. The function
returns an object of the S3 class "linreg" which can be explored with the following methods:
\itemize{
  \item \code{\link{coef.linreg}}
  \item \code{\link{plot.linreg}}
  \item \code{\link{pred}}
  \item \code{\link{print.linreg}}
  \item \code{\link{residuals.linreg}}
  \item \code{\link{summary.linreg}}
}
}
\details{
The second function is called \code{\link{ridgereg}}. This function performs ridge regression using ordinary least squares. The function
returns an object of the S3 class "ridgereg" which can be explored with the following methods:
\itemize{
  \item \code{\link{coef.ridgereg}}
  \item \code{\link{predict.ridgereg}}
  \item \code{\link{print.ridgereg}}
}
}
