% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{Perform linear regression with OLS}
\usage{
linreg(formula, data)
}
\arguments{
\item{formula}{An object of the class \code{\link[stats]{formula}}.}

\item{data}{A data frame.}
}
\description{
Perform linear regression with OLS
}
