% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.linreg}
\alias{plot.linreg}
\title{Plot method for linreg objects}
\usage{
\method{plot}{linreg}(object, ...)
}
\arguments{
\item{object}{an object of class "linreg", generated by a call to \code{\link{linreg}}.}
}
\value{
Two scatter plots. The first one is residuals vs fitted values and the second is the scale-location.
}
\description{
\code{plot} method for objects of the class .
}
