% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{predict.linreg}
\alias{predict.linreg}
\title{Predict method for linreg objects}
\usage{
\method{predict}{linreg}(object, ...)
}
\arguments{
\item{object}{An object of class "linreg", generated by a call to \code{\link{linreg}}.}
}
\value{
A vector of predicted values.
}
\description{
Predicted values from an object of the class "linreg".
}
