% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.linreg}
\alias{print.linreg}
\title{Print out the estimated coeffiencts}
\usage{
\method{print}{linreg}(object, ...)
}
\arguments{
\item{object}{An object of Class "linreg", generated by a call to \code{\link{linreg}}.}

\item{...}{additional arguments.}
}
\value{
A list of estimated coeffiencts.
}
\description{
Estimated coeffiencts from an object of class "linreg".
}
