% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.linreg}
\alias{print.linreg}
\title{Print out the estimated coeffiencts}
\usage{
\method{print}{linreg}(x)
}
\arguments{
\item{object}{An object of Class}
}
\value{
A list of estimated coeffiencts
}
\description{
Estimated coeffiencts from an object of class
}
